LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/forms/select-multiple-scrolling.optional-expected.txt



FAIL select[multiple][style="writing-mode: horizontal-tb"] scrolls correctly assert_equals: Setting scrollTop to a negative value of the list does not work for writing-mode: horizontal-tb expected 0 but got 98
FAIL select[multiple][style="writing-mode: vertical-lr"] scrolls correctly assert_true: select should be scrollable in block axis expected true got false
FAIL select[multiple][style="writing-mode: vertical-rl"] scrolls correctly assert_true: select should be scrollable in block axis expected true got false
